NuGet Package (#111)
Created new NuGet package. Removed need to manually update the Sqlite DB. Refactored naming of FileRef to DocumentReference to more clearly describe an artifact that could be a file, web page, or other form of document that might not be associated with a file system. NPOI library stopped working on *.doc files, so I removed support and opened a new issue.

Closes #108