Chat History Updates (#144)
Since we want to work with both user and assistant messages, this changes the history type to their ChatRequestMessage base class. Since this base class only lets us read the Role, I added a utility method to AzureOpenAIChatRequest to get the content based on the request type. Also removed some history management logic, instead leaving that to the calling code to decide how history is managed.

Closes #143