Fixed Config Folder Connection. (#153)
There was a problem where setting the `Generellem.SubFolder` property resulted in a SQLite exception. The reason is that SQLite isn't able to create new folders on it's own. So, I put that check in where I get the path for configuration to check and create that folder if it isn't there.